+++ machine_learning/exter_dataset/uitls/download.py
@@ -20,7 +20,6 @@ except:
     api =None
 
 def zip_download(base_path,dataset_name,url,name=None):
-    print("zip_download:",dataset_name)
     if name is not None:
         dir_name = os.path.join(base_path,"repo",dataset_name)
         zip_name = os.path.join(base_path,"repo",dataset_name,name+".zip")
@@ -28,20 +27,18 @@ def zip_download(base_path,dataset_name,url,name=None):
         zip_name = os.path.join(base_path,"repo",dataset_name,dataset_name+".zip")
         dir_name = os.path.join(base_path,"repo",dataset_name)
     
-    if not os.path.isdir(dir_name):
-        os.mkdir(dir_name)
-    try:
-        dir = os.listdir(dir_name)
-        if len(dir) != 0:
+    if os.path.isdir(dir_name) and os.path.exists(dir_name):
+        if len(os.listdir(dir_name)) != 0:
             return
-    except:
-        pass
+    else:
+        os.mkdir(dir_name)
     
     if not os.path.exists(zip_name):
         with urllib.request.urlopen(url) as f:
             text = f.read()
             with open(zip_name,"wb") as file :
                 file.write(text)
+    
     try:
         with zipfile.ZipFile(zip_name,"r") as zip_ref:
             zip_ref.extractall(dir_name)
@@ -51,7 +48,6 @@ def zip_download(base_path,dataset_name,url,name=None):
 
 
 def gz_tar_download(base_path,dataset_name,url,name=None):
-    print("gz_tar_download:",dataset_name)
     if name is not None:
         dir_name = os.path.join(base_path,"repo",dataset_name)
         zip_name = os.path.join(base_path,"repo",dataset_name,dataset_name+".gz_tar")
@@ -76,35 +72,26 @@ def gz_tar_download(base_path,dataset_name,url,name=None):
         zip_ref.extractall(path=dir_name)
 
 def git_download(base_path,dataset_name,url):
-    print("git_download:",dataset_name)
+    
     dir_name = os.path.join(base_path,"repo",dataset_name)
-    if not os.path.isdir(dir_name):
-        os.mkdir(dir_name)
-    dir = os.listdir(dir_name)
-    if len(dir) == 0:
-        try:
-            Repo.clone_from(url, dir_name)
-        except:
-            print("fail git_download:",dataset_name)
+    
+    
+    if os.path.isdir(dir_name) and os.path.exists(dir_name):
+        if len(os.listdir(dir_name)) != 0:
+            return
     else:
-        try:
-            g = git.Git(dir_name)
-            o = g.remotes.origin
-            o.pull()
-        except:
-            print("fail update git:",dataset_name)
-
+        os.mkdir(dir_name)
+    print("download git")
+    Repo.clone_from(url, dir_name)
 
 def kaggle_download(base_path,dataset_name,name):
-    print("kaggle_download:",dataset_name)
     path = os.path.join(base_path,"repo",dataset_name, name)
     try:
         api.dataset_download_files(name,path=path, unzip=True)
     except:
-        print("fail kaggle_download:",dataset_name)
+        pass
 
 def huggingface_download(base_path,dataset_name,name,subname):
-    print("huggingface_download:",dataset_name)
     dir_fs = os.path.join(base_path,"repo", dataset_name)
     if not os.path.isdir(dir_fs):
         os.mkdir(dir_fs)
@@ -115,7 +102,6 @@ def huggingface_download(base_path,dataset_name,name,subname):
     return dataset
 
 def zenodo_download(base_path,dataset_name,id_zenodo):
-    print("zenodo_download:",dataset_name)
     from zenodo_client import Zenodo
     dir_fs = os.path.join(base_path,"repo", dataset_name)
     
@@ -129,11 +115,10 @@ def zenodo_download(base_path,dataset_name,id_zenodo):
         zenodo = Zenodo()
         dataset = zenodo.download_latest(id_zenodo, path=dir_fs)
     except:
-        print("fail zenodo_download:",dataset_name)
+        pass
     return dataset
 
 def drive_google_download(base_path,dataset_name,id_google_dive):
-    print("drive_google_download")
     pass
 def onedrive_download(base_path,dataset_name,id_zenodo):
     pass
@@ -144,7 +129,6 @@ def mumin_download(base_path,dataset_name,size):
     dataset = MuminDataset(bearer_token, size='small')
 
 def file_download(base_path,dataset_name,url,name=None,ext=None):
-    print("file_download:",dataset_name)
     dir_name = os.path.join(base_path,"repo",dataset_name)
     if name is not None:
         path_name = os.path.join(base_path,"repo",dataset_name, name)
@@ -152,17 +136,9 @@ def file_download(base_path,dataset_name,url,name=None,ext=None):
         path_name = os.path.join(base_path,"repo",dataset_name,dataset_name+"."+ext)
     else:
         path_name = os.path.join(base_path,"repo",dataset_name,dataset_name)
-    
-    try:
-        dir = os.listdir(dir_name)
-        if len(dir) != 0:
-            return
-    except:
-        pass
-    
     if not os.path.isdir(dir_name):
         os.mkdir(dir_name)
-    if not os.path.exists(path_name):
+    if not os.path.isdir(path_name) and os.path.exists(path_name):
         with urllib.request.urlopen(url) as f:
             text = f.read()
             with open(path_name,mode="wb") as file :
@@ -184,11 +160,11 @@ def dropbox_download(base_path,dataset_name,url):
 
 
 def kaggle_download(base_path,user,dataset_name,file_name):
-    print("kaggle_download",dataset_name)
     dir_name = os.path.join(base_path,"repo",dataset_name)
     if not os.path.isdir(dir_name):
         os.mkdir(dir_name)
     dir_name = os.path.join(base_path,"repo",dataset_name,file_name)
-    text= api.datasets_download_file(user,dataset_name,file_name=file_name)
-    with open(dir_name,mode="wb") as file :
-            file.write(text)
+    if  os.path.isdir(dir_name) and os.path.exists(dir_name):
+        text= api.datasets_download_file(user,dataset_name,file_name=file_name)
+        with open(dir_name,mode="wb") as file :
+                file.write(text)
